move embedding code from modules to functional
