+++ python/oneflow/nn/functional/__init__.py
@@ -82,7 +82,7 @@ from oneflow._C import binary_cross_entropy_loss as binary_cross_entropy
 from oneflow._C import (
     binary_cross_entropy_with_logits_loss as binary_cross_entropy_with_logits,
 )
-from oneflow.nn.modules.sparse import embedding
+from .sparse import embedding
 from oneflow.nn.modules.linear import linear
 from oneflow.nn.modules.activation import relu6
 from oneflow.nn.modules.upsampling import Upsample as upsample
